refactor: e2k関連処理をファイル分割
